lora_trainer_suite/requirements.txt
# LoRA Trainer Suite Requirements
# Optimized for RTX 3090, 128GB RAM, Ryzen 9 7900X

# Core Dependencies
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Diffusion Models
diffusers>=0.25.0
transformers>=4.37.0
accelerate>=0.25.0
safetensors>=0.4.0
peft>=0.8.0

# Quantization & Optimization
bitsandbytes>=0.41.0
flash-attn>=2.5.0
xformers>=0.0.23
triton>=2.1.0

# CLIP Interrogator
clip-interrogator>=0.6.0
open_clip_torch>=2.23.0

# Vision-Language Models (Qwen2-VL)
qwen-vl-utils>=0.0.1
einops>=0.7.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.1

# GUI
gradio>=4.15.0
plotly>=5.18.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0.1

# Training Frameworks (optional - install separately)
# kohya-ss - install from https://github.com/kohya-ss/sd-scripts
# ai-toolkit - install from https://github.com/ostris/ai-toolkit

# Utilities
rich>=13.7.0
colorama>=0.4.6
python-dotenv>=1.0.0

# Development (optional)
pytest>=7.4.0
black>=23.12.0
flake8>=6.1.0

# Optional: WandB for experiment tracking
# wandb>=0.16.0

# Optional: Tensorboard
# tensorboard>=2.15.0
# tensorboardX>=2.6.0

# Note: For Qwen2-VL, you may need to install from source or specific repo:
# pip install git+https://github.com/QwenLM/Qwen2-VL.git
